client_service/requirements.txt
python-socketio
requests
websocket-client
pyttsx3
sounddevice
soundfile
numpy